Match scene-specific BGM helpers in C

+++ include/field_bgm.h
@@ -15,7 +15,7 @@ u16 FieldBGM_GetEffective(FieldSystem *fieldSystem, u32 mapId);
 u16 FieldBGM_GetForMapHeader(FieldSystem *fieldSystem, int mapId);
 BOOL FieldBGM_TryFadeOut(FieldSystem *fieldSystem, u16 seqNo, int mode);
 static void FieldBGM_GetFadeOutAndWaitFrames(FieldSystem *fieldSystem, int mode, int *fadeOutFrames, int *waitFrames);
-int FieldBGM_GetEyesMeetForTrainer(u16 a0, int regionNo);
+u16 FieldBGM_GetEyesMeetForTrainer(u16 a0, int regionNo);
 void FieldBGM_TryFadeIn(FieldSystem *fieldSystem, u32 mapId);
 void FieldBGM_PlayForMapHeader(FieldSystem *fieldSystem, u32 mapId, BOOL a2);
 BOOL FieldBGM_PlayEffectiveForMapHeader(FieldSystem *fieldSystem, u32 mapId);
